Read out environment from nix-shell and cache it
This fixes issue #21, the massive disk space leak due to successive
calls to nix-shell. The patch reads out the complete environment
of a nix sandbox with `nix-shell --run 'printenv -0' 2> /dev/null`
and writes it into an temporary file for later usage. If the user
now wants to execute a command inside the sandbox, instead of using
`nix-sandbox --run`, it's execution is performed with
`sh -c 'source /tmp/nix-sandbox-rc; command'`.